refactor: extract summary helper into testable module
